feat: :sparkles:
Adds Hero, About, Features, and FAQ sections to the landing page. Also, refactors page.tsx to use the new components.
