+++ app/components/index.ts
@@ -1,0 +1,4 @@
+export { FeaturesSection } from '@/app/components/FeaturesSection';
+export { AboutSection } from '@/app/components/AboutSection';
+export { HeroSection } from "@/app/components/HeroSection";
+export { FAQSection } from '@/app/components/FAQSection';
